Refactor query parameter handling: replace appendQueryParams utility with useAppendQueryParams hook across components. Update Link, CallToAction, and Hero components to utilize the new hook for appending query parameters. Remove old utility function from utils.ts and add new hook in useQueryParams.ts.
